+++ app/src/main/java/com/meong/podoandroid/ui/feed/FeedRecommendAdapter.java
@@ -1,9 +1,7 @@
-package com.meong.podoandroid;
+package com.meong.podoandroid.ui.feed;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.provider.ContactsContract;
-import android.text.Layout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,6 +13,8 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.meong.podoandroid.R;
+import com.meong.podoandroid.data.FeedData;
 
 import java.util.ArrayList;
