+++ app/src/main/java/com/meong/podoandroid/ui/map/MapSearchActivity.java
@@ -1,4 +1,4 @@
-package com.meong.podoandroid;
+package com.meong.podoandroid.ui.map;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.drawerlayout.widget.DrawerLayout;
@@ -10,16 +10,18 @@ import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-public class MainActivity extends AppCompatActivity {
+import com.meong.podoandroid.R;
+import com.meong.podoandroid.ui.menu.MainActivity;
 
-    ImageView imgMenu;
+public class MapSearchActivity extends AppCompatActivity {
+
     DrawerLayout drawer;
+    ImageView imgMenu;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
+        setContentView(R.layout.activity_map_search);
 
         setDrawer();
         setOnBtnClickListener();
@@ -39,35 +41,15 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        //병원 위치가 눌렸을 때
-        TextView txtHospital = findViewById(R.id.txt_nav_main_hospital);
-        txtHospital.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), MapSearchActivity.class);
-                startActivity(intent);
-
-                finish();
-            }
-        });
-
-        //사료 추천이 눌렸을 때
-        TextView txtRecommend = findViewById(R.id.txt_nav_main_recommend);
-        txtRecommend.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-
-            }
-        });
     }
 
     private void setDrawer() {
-        drawer = (DrawerLayout) findViewById(R.id.drawer_main_act);
+        drawer = (DrawerLayout) findViewById(R.id.drawer_map_search_act);
         drawer.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
     }
 
     private void setOnBtnClickListener() {
-        imgMenu = (ImageView) findViewById(R.id.img_main_act_menu);
+        imgMenu = (ImageView) findViewById(R.id.map_hamburger);
         imgMenu.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
