+++ app/src/main/res/layout/activity_feed_recommend.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#dee9ff"
-    tools:context=".FeedRecommendActivity">
+    tools:context=".ui.feed.FeedRecommendActivity">
     
     <RelativeLayout
         android:id="@+id/relativelayout"
